requirements.txt
astroid==2.1.0
autopep8==1.4.3
backcall==0.1.0
certifi==2018.11.29
colorama==0.4.1
decorator==4.3.0
Django==2.1.3
django-crispy-forms==1.7.2
django-extensions==2.1.4
django-filter==2.0.0
django-guardian==1.4.9
django-heroku==0.3.1
djangorestframework==3.9.0
docutils==0.14
graphviz==0.10.1
gunicorn==19.9.0
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
parso==0.3.1
pickleshare==0.7.5
prompt-toolkit==2.0.6
Pygments==2.3.0
pytz==2018.7
setuptools==40.6.2
simplegeneric==0.8.1
traitlets==4.3.2
wcwidth==0.1.7